+++ checkpoint/CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - #v1 Add `use_load_and_broadcast` option.
 - Add PyTorch DCP (Distributed Checkpoint) to the benchmark suite.
+- #v1 #safetensors Implement `load_pytree_async`, allowed only for the
+Safetensors format at the moment.
 - #v1 Add `DeletionOptions` to configure V1 Checkpointer's checkpoint deletion
 behavior.
 - #v1 Add `cleanup_tmp_directories` setting to V1 Checkpointer to manage